Task #1418: fix various MAC protocols to restore compat with 64 bit Linux (size_t -> uint32). Most of this stuff elsewhere was already patched earlier by Ruud, and some supplemental fixes by me, but apparently there is more and this stuff is not in jenkins(?). The LOFAR tree now appears to build (I do that without PVSS).

+++ MAC/APL/CR_Protocol/src/CRrecordRequest.cc
@@ -47,7 +47,7 @@ size_t CRrecordRequest::getSize()
 
 size_t CRrecordRequest::pack  (char* buffer) const
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 	MSH_pack(buffer, offset, stationList);	
 	MSH_pack(buffer, offset, rcuList);	
 	return (offset);
@@ -55,7 +55,7 @@ size_t CRrecordRequest::pack  (char* buffer) const
 
 size_t CRrecordRequest::unpack(const char *buffer)
 {
-	size_t offset = 0;
+	uint32 offset = 0;
 	MSH_unpack(buffer, offset, stationList);	
 	MSH_unpack(buffer, offset, rcuList);	
 	return (offset);
